+++ out/commands.js
@@ -6,6 +6,7 @@ const path = require("path");
 const extension_config = vscode.workspace.getConfiguration(this.id);
 const isWin = process.platform === "win32";
 const language = require("./language_defines");
+require("./utils");
 
 class CommandsManager {
     constructor() {
@@ -28,7 +29,7 @@ class CommandsManager {
         let porth_path_conf = extension_config.get('porth.path');
         let porth_path = porth_path_conf != "_builtin_" ? porth_path_conf : path.join(context.extensionPath, "/porth");
 
-        if (this.open_file_name == null || this.open_file_path == null) {
+        if ((this.open_file_name == null || this.open_file_path == null) && action != language.CMD.BOOTSTRAP) {
             vscode.window.showErrorMessage("[Porth]: Please open a .porth file first");
             console.log("Couldn't select a visible editor containing a .porth file");
             return;
@@ -40,11 +41,23 @@ class CommandsManager {
         let flag_debug = extension_config.get('porth.debug');
         console.log(`Using global extension settings: [Autorun: ${flag_autorun}, Debug: ${flag_debug}]`);
 
+        fs.stat(path.join(porth_path, "/porth"), (err, stats) => {
+            if (err && err.code === 'ENOENT') {
+                console.log(`Couldn't find porth compiler in ${porth_path}.`);
+                vscode.window.showErrorMessage("[Porth]: Compiler not found. Bootstrap before compiling.");
+                return;
+            }
+            else if (err) {
+                vscode.window.showErrorMessage('[Porth]: Error while checking if compiler exists: ' + err);
+                return;
+            }
+        });
+
         this.buildCommand(action, porth_path, flag_autorun, flag_debug);
     }
 
     /**
-     *  @brief  Assembles the command to be executed depending on the action
+     *  @brief  Assembles the command to be executed based on the action and os
      *  @param  String action
      *  @param  String porth_path
      *  @param  Boolean flag_autorun
@@ -55,22 +68,52 @@ class CommandsManager {
         let cmd = "";
         let args = [];
         switch (action) {
-            // case language.CMD.SIMULATE:
             case language.CMD.COMPILE:
-                cmd = `python3`;
-                args = [`${path.join(porth_path, "/porth.py")}`, `-I`, `${porth_path}/std`];
-                if (flag_debug) args.push("-debug");
-                args.push(`${action}`);
-                if (flag_autorun && action == language.CMD.COMPILE) args.push("-r");
-                args.push(`${this.open_file_path}`);
-                return this.executeCommand(cmd, args);
+                if (isWin) {
+                    cmd = `cd ${porth_path} && wsl`;
+                    args = [`${path.posix.join(porth_path.winToWslPath(), "/porth")}`];
+                    // if (flag_debug) args.push("-debug");
+                    args.push(`${action}`);
+                    if (flag_autorun) args.push("-r");
+                    args.push(`${this.open_file_path.winToWslPath()}`);
+                    this.executeCommand(cmd, args);
+                } else {
+                    cmd = `cd ${porth_path} && ${path.posix.join(porth_path, "/porth")}`;
+                    args = [];
+                    // if (flag_debug) args.push("-debug");
+                    args.push(`${action}`);
+                    if (flag_autorun) args.push("-r");
+                    args.push(`${this.open_file_path}`);
+                    this.executeCommand(cmd, args);
+                }
+                break;
+            case language.CMD.BOOTSTRAP:
+                cmd = "nasm";
+                args = [`-felf64`, `${path.posix.join(porth_path, "/bootstrap/porth-linux-x86_64.nasm")}`, `-o`, `${path.posix.join(porth_path, "/porth.o")}`];
+                this.executeCommand(cmd, args);
+                if (isWin) {
+                    setTimeout(() => {
+                        cmd = "wsl";
+                        args = [`ld`, `-o`, `${path.posix.join(porth_path.winToWslPath(), "/porth")}`, `${path.posix.join(porth_path.winToWslPath(), "/porth.o")}`];
+                        args.push(...[`&&`, `cd`, `${porth_path}`, `&&`, `wsl`, `${path.posix.join(porth_path.winToWslPath(), "/porth")}`, `com`, `${path.posix.join(porth_path.winToWslPath(), "/porth.porth")}`]);
+                        this.executeCommand(cmd, args);
+                    }, 1000);
+                } else {
+                    setTimeout(() => {
+                        cmd = "ld";
+                        args = [`-o`, `${path.posix.join(porth_path, "/porth")}`, `${path.posix.join(porth_path, "/porth.o")}`];
+                        args.push(...[`${path.posix.join(porth_path, "/porth")}`, `com`, `${path.posix.join(porth_path, "/porth.porth")}`]);
+                        this.executeCommand(cmd, args);
+                    }, 1000);
+                }
+                break;
             case language.CMD.RUN:
                 let basepath = this.open_file_path.split(".porth")[0];
                 fs.stat(basepath, (err, stats) => {
                     if (err && err.code === 'ENOENT') {
                         console.log(`${basepath} does not exist.`);
                         vscode.window.showErrorMessage("[Porth]: Executable not found. Compile before running.");
-                        return this.executeCommand(cmd, args);
+                        return;
                     }
                     else if (err) {
                         vscode.window.showErrorMessage('[Porth]: Error while checking if output file exists: ' + err);
@@ -78,12 +121,12 @@ class CommandsManager {
                     } else {
                         if (isWin) {
                             cmd = `wsl`;
-                            args = [`${basepath.replace(/\\/g, '/').replace(/(^)/, '/mnt/$1').replace(':', '')}`];
+                            args = [`${basepath.winToWslPath()}`];
                         } else {
                             cmd = `${basepath}`;
                             args = [];
                         }
-                        console.log(`Selected ${basepath} as the executable to run.`);
+                        console.log(`Running execuable: ${basepath}`);
                         return this.executeCommand(cmd, args);
                     }
                 });
@@ -98,10 +141,12 @@ class CommandsManager {
     executeCommand = (cmd, args) => {
         console.log(`Executing command [${cmd}] with args [${args}]...`);
         if (cmd == "") return;
+        let terminal;
         if (vscode.window.activeTerminal != undefined && vscode.window.activeTerminal.name == "Porth") {
-            vscode.window.activeTerminal.dispose();
+            terminal = vscode.window.activeTerminal;
+        } else {
+            terminal = vscode.window.createTerminal("Porth");
         }
-        let terminal = vscode.window.createTerminal("Porth");
 
         terminal.show();