+++ out/extension.js
@@ -37,11 +37,11 @@ function activate(context) {
         }
     }
 
-    // let simulate = vscode.commands.registerCommand('porth.simulate', () => {
-    //     commandsManager.prepareCommand(language.CMD.SIMULATE, context);
-    // });
     let compile = vscode.commands.registerCommand('porth.compile', () => {
         commandsManager.prepareCommand(language.CMD.COMPILE, context);
+    });
+    let bootstrap = vscode.commands.registerCommand('porth.bootstrap', () => {
+        commandsManager.prepareCommand(language.CMD.BOOTSTRAP, context);
     });
     let run = vscode.commands.registerCommand('porth.run', () => {
         commandsManager.prepareCommand(language.CMD.RUN, context);
@@ -75,7 +75,7 @@ function activate(context) {
         }
     });
 
-    context.subscriptions.push(compile, run, open_documentation, hoverProvider, definitionProvider);
+    context.subscriptions.push(compile, bootstrap, run, open_documentation, hoverProvider, definitionProvider);
 }
 
 exports.activate = activate;